backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
mysql-connector-python==8.1.0
python-dotenv==1.0.0
requests==2.31.0
transformers==4.35.0
torch==2.8.0
sqlparse==0.4.4
nltk==3.8.1